Fixed parsing of ArrayList during request bulding
There was some extra work done badly when the object being processed during request build time was an ArrayList. The implementation expected the contents of the array to be a Map and then iterated over it. Instead, now every element of the List is processed and the parseParameters function is called recursively with the List element as input, this way we let the recursive function to go down iterating over the collections until it gets to the primary elements